+++ src/checks.rs
@@ -132,7 +132,7 @@ mod tests {
     fn run_with_invalid_line_test() {
         let line = line_entry(1, 2, "FOO");
         let warning = Warning::new(
-            line.clone(),
+            line.number,
             LintKind::KeyWithoutValue,
             "The FOO key should be with a value or have an equal sign",
         );
@@ -147,7 +147,7 @@ mod tests {
     fn run_without_blank_line_test() {
         let line = line_entry(1, 1, "FOO=BAR");
         let warning = Warning::new(
-            line.clone(),
+            line.number,
             LintKind::EndingBlankLine,
             "No blank line at the end of the file",
         );
@@ -163,7 +163,7 @@ mod tests {
         let line1 = line_entry(1, 3, "FOO\n");
         let line2 = line_entry(2, 3, "1FOO\n");
         let warning = Warning::new(
-            line2.clone(),
+            line2.number,
             LintKind::LeadingCharacter,
             "Invalid leading character detected",
         );
@@ -190,7 +190,7 @@ mod tests {
         let line2 = line_entry(2, 4, "FOO\n");
         let line3 = line_entry(3, 4, "1FOO\n");
         let warning = Warning::new(
-            line3.clone(),
+            line3.number,
             LintKind::LeadingCharacter,
             "Invalid leading character detected",
         );
@@ -207,7 +207,7 @@ mod tests {
         let line2 = line_entry(2, 4, "FOO\n");
         let line3 = line_entry(3, 4, "1FOO\n");
         let warning = Warning::new(
-            line3.clone(),
+            line3.number,
             LintKind::LeadingCharacter,
             "Invalid leading character detected",
         );
@@ -229,7 +229,7 @@ mod tests {
         let line3 = line_entry(3, 5, "# dotenv-linter:on LeadingCharacter\n");
         let line4 = line_entry(4, 5, "1FOO\n");
         let warning = Warning::new(
-            line4.clone(),
+            line4.number,
             LintKind::LeadingCharacter,
             "Invalid leading character detected",
         );
@@ -244,7 +244,7 @@ mod tests {
     fn only_simple_comment() {
         let line = line_entry(1, 1, "# Simple comment");
         let warning = Warning::new(
-            line.clone(),
+            line.number,
             LintKind::EndingBlankLine,
             "No blank line at the end of the file",
         );