+++ src/checks/ending_blank_line.rs
@@ -22,7 +22,7 @@ impl Default for EndingBlankLineChecker<'_> {
 impl Check for EndingBlankLineChecker<'_> {
     fn run(&mut self, line: &LineEntry) -> Option<Warning> {
         if line.is_last_line() && !line.raw_string.ends_with(LF) {
-            Some(Warning::new(line.clone(), self.name(), self.message()))
+            Some(Warning::new(line.number, self.name(), self.message()))
         } else {
             None
         }