+++ src/fixes/lowercase_key.rs
@@ -14,7 +14,7 @@ impl Fix for LowercaseKeyFixer {
         LintKind::LowercaseKey
     }
 
-    fn fix_line(&mut self, line: &mut LineEntry) -> Option<()> {
+    fn fix_line(&self, line: &mut LineEntry) -> Option<()> {
         let key = line.get_key()?;
         let key = key.to_uppercase();
         line.raw_string = format!("{}={}", key, line.get_value()?);
@@ -26,11 +26,11 @@ impl Fix for LowercaseKeyFixer {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::common::{tests::*, Warning};
+    use crate::common::tests::*;
 
     #[test]
     fn fix_line_test() {
-        let mut fixer = LowercaseKeyFixer::default();
+        let fixer = LowercaseKeyFixer::default();
         let mut line = line_entry(1, 1, "foO=BAR");
 
         assert_eq!(Some(()), fixer.fix_line(&mut line));
@@ -39,19 +39,15 @@ mod tests {
 
     #[test]
     fn fix_warnings_test() {
-        let mut fixer = LowercaseKeyFixer::default();
+        let fixer = LowercaseKeyFixer::default();
         let mut lines = vec![
             line_entry(1, 3, "foO=BAR"),
             line_entry(2, 3, "Z=Y"),
             blank_line_entry(3, 3),
         ];
-        let mut warning = Warning::new(
-            lines[0].clone(),
-            LintKind::LowercaseKey,
-            String::from("The FOO key should be in uppercase"),
-        );
+        let warning_lines = [lines[0].number];
 
-        assert_eq!(Some(1), fixer.fix_warnings(vec![&mut warning], &mut lines));
+        assert_eq!(Some(1), fixer.fix_warnings(&warning_lines, &mut lines));
         assert_eq!("FOO=BAR", lines[0].raw_string);
     }
 }