+++ src/fixes/quote_character.rs
@@ -14,7 +14,7 @@ impl Fix for QuoteCharacterFixer {
         LintKind::QuoteCharacter
     }
 
-    fn fix_line(&mut self, line: &mut LineEntry) -> Option<()> {
+    fn fix_line(&self, line: &mut LineEntry) -> Option<()> {
         let value = line.get_value()?;
         let pure_val = value.replace("'", "").replace("\"", "");
 
@@ -27,11 +27,11 @@ impl Fix for QuoteCharacterFixer {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::common::{tests::*, Warning};
+    use crate::common::tests::*;
 
     #[test]
     fn fix_line_test() {
-        let mut fixer = QuoteCharacterFixer::default();
+        let fixer = QuoteCharacterFixer::default();
         let mut line = line_entry(1, 1, "FOO=\'\"b\'\"ar\"\'");
 
         assert_eq!(Some(()), fixer.fix_line(&mut line));
@@ -40,19 +40,15 @@ mod tests {
 
     #[test]
     fn fix_warnings_test() {
-        let mut fixer = QuoteCharacterFixer::default();
+        let fixer = QuoteCharacterFixer::default();
         let mut lines = vec![
             line_entry(1, 3, "FOO=\"bar\'\""),
             line_entry(2, 3, "Z=Y"),
             blank_line_entry(3, 3),
         ];
-        let mut warning = Warning::new(
-            lines[0].clone(),
-            LintKind::QuoteCharacter,
-            "The value has quote characters (\', \")",
-        );
+        let warning_lines = [lines[0].number];
 
-        assert_eq!(Some(1), fixer.fix_warnings(vec![&mut warning], &mut lines));
+        assert_eq!(Some(1), fixer.fix_warnings(&warning_lines, &mut lines));
         assert_eq!("FOO=bar", lines[0].raw_string);
     }
 }