+++ src/fixes/trailing_whitespace.rs
@@ -14,7 +14,7 @@ impl Fix for TrailingWhitespaceFixer {
         LintKind::TrailingWhitespace
     }
 
-    fn fix_line(&mut self, line: &mut LineEntry) -> Option<()> {
+    fn fix_line(&self, line: &mut LineEntry) -> Option<()> {
         line.raw_string = line.raw_string.trim_end().to_string();
 
         Some(())
@@ -24,11 +24,11 @@ impl Fix for TrailingWhitespaceFixer {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::common::{tests::*, Warning};
+    use crate::common::tests::*;
 
     #[test]
     fn fix_line_test() {
-        let mut fixer = TrailingWhitespaceFixer::default();
+        let fixer = TrailingWhitespaceFixer::default();
         let mut line = line_entry(1, 1, "DEBUG_HTTP=true  ");
 
         assert_eq!(Some(()), fixer.fix_line(&mut line));
@@ -37,19 +37,15 @@ mod tests {
 
     #[test]
     fn fix_warnings_test() {
-        let mut fixer = TrailingWhitespaceFixer::default();
+        let fixer = TrailingWhitespaceFixer::default();
         let mut lines = vec![
             line_entry(1, 3, "FOO=BAR "),
             line_entry(2, 3, "Z=Y"),
             blank_line_entry(3, 3),
         ];
-        let mut warning = Warning::new(
-            lines[0].clone(),
-            LintKind::TrailingWhitespace,
-            "Trailing whitespace detected",
-        );
+        let warning_lines = [lines[0].number];
 
-        assert_eq!(Some(1), fixer.fix_warnings(vec![&mut warning], &mut lines));
+        assert_eq!(Some(1), fixer.fix_warnings(&warning_lines, &mut lines));
         assert_eq!("FOO=BAR", lines[0].raw_string);
     }
 }