+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 🚀 Added
 
 ### 🔧 Changed
+- Replace `LineEntry` with line number in `Warning` [#447](https://github.com/dotenv-linter/dotenv-linter/pull/447) ([@mgrachev](https://github.com/mgrachev))
 - Tidy up imports [#446](https://github.com/dotenv-linter/dotenv-linter/pull/446) ([@mgrachev](https://github.com/mgrachev))
 - Add type alias for `Result` [#445](https://github.com/dotenv-linter/dotenv-linter/pull/445) ([@mgrachev](https://github.com/mgrachev))
 - Change edition to 2021 [#444](https://github.com/dotenv-linter/dotenv-linter/pull/444) ([@mgrachev](https://github.com/mgrachev))